requirements.txt
Django==2.0.7
graphene-django==2.0.0
django-filter==1.1.0
django-graphql-jwt==0.1.10
